add tool to report cmap coverage across noto fonts
